pyproject.toml
[project]
name = "dotmate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.5",
    "pydantic>=2.8.2",
    "types-requests>=2.32.0.20240914",
    "pyyaml>=6.0.2",
    "apscheduler>=3.10.4",
]
